+++ custom_components/spotifyplus/config_flow.py
@@ -18,15 +18,28 @@ from __future__ import annotations
 from collections.abc import Mapping
 import logging
 from typing import Any
+import voluptuous as vol
 
 from spotifywebapipython import SpotifyClient
+from spotifywebapipython.models import Device
 
-from homeassistant.config_entries import ConfigEntry
+from homeassistant.config_entries import ConfigEntry, OptionsFlow
 from homeassistant.const import CONF_DESCRIPTION, CONF_ID, CONF_NAME
+from homeassistant.core import callback
 from homeassistant.data_entry_flow import FlowResult
 from homeassistant.helpers import config_entry_oauth2_flow
+from homeassistant.helpers.selector import (
+    SelectSelector,
+    SelectSelectorConfig,
+    SelectSelectorMode,
+)
 
-from .const import DOMAIN, SPOTIFY_SCOPES
+from .const import (
+    CONF_OPTION_DEVICE_DEFAULT, 
+    DOMAIN, 
+    SPOTIFY_SCOPES
+)
+from .instancedata_spotifyplus import InstanceDataSpotifyPlus
 
 # get smartinspect logger reference; create a new session for this module name.
 from smartinspectpython.siauto import SIAuto, SILevel, SISession, SIColors
@@ -240,6 +253,193 @@ class SpotifyFlowHandler(config_entry_oauth2_flow.AbstractOAuth2FlowHandler, dom
             # trace.
             _logsi.LogException(None, ex, logToSystemLogger=False)
             raise
+        
+        finally:
+
+            # trace.
+            _logsi.LeaveMethod(SILevel.Debug)
+
+    @staticmethod
+    @callback
+    def async_get_options_flow(config_entry: ConfigEntry) -> OptionsFlow:
+        """
+        Get the options flow for this handler, which enables options support.
+
+        This method is invoked when a user clicks the "Configure" button from the integration
+        details page of the UI.
+        """
+        return SpotifyPlusOptionsFlow(config_entry)
+
+
+class SpotifyPlusOptionsFlow(OptionsFlow):
+    """
+    Handles options flow for the component.
+    
+    The options flow allows a user to configure additional options for the component at any time by 
+    navigating to the integrations page and clicking the Options button on the card for your component. 
+    Generally speaking these configuration values are optional, whereas values in the config flow are 
+    required to make the component function.
+    """
+
+    def __init__(self, entry:ConfigEntry) -> None:
+        """
+        Initialize options flow.
+        """
+        try:
+
+            # trace.
+            _logsi.EnterMethod(SILevel.Debug)
+            _logsi.LogObject(SILevel.Verbose, "'%s': OptionsFlow is initializing - entry (ConfigEntry) object" % entry.title, entry)
+            _logsi.LogDictionary(SILevel.Verbose, "'%s': OptionsFlow entry.data dictionary" % entry.title, entry.data)
+            _logsi.LogDictionary(SILevel.Verbose, "'%s': OptionsFlow entry.options dictionary" % entry.title, entry.options)
+       
+            # initialize storage.
+            self._entry = entry
+            self._name:str = None
+
+            # load config entry base values.
+            self._name = entry.data.get(CONF_NAME, None)
+
+            # load config entry options values.
+            self._Options = dict(entry.options)
+            
+        except Exception as ex:
+            
+            # trace.
+            _logsi.LogException(None, ex, logToSystemLogger=False)
+            raise
+        
+        finally:
+
+            # trace.
+            _logsi.LeaveMethod(SILevel.Debug)
+
+
+    async def async_step_init(self, user_input:dict[str,Any]=None) -> FlowResult:
+        """
+        Manage the options for the custom component.
+
+        Args:
+            user_input (dict[str,Any]):
+                User input gathered from the input form.  
+                This argument defaults to None when this step is first called.  
+                When the user clicks the submit button on the form, the argument will contain
+                a dictionary of the data that was entered.  Home Assistant will do some basic 
+                # validation on your behalf based on the data schema that you defined (e.g. 
+                # required field, port number is within a numeric range, etc). 
+                
+        For a good example, look at HA demo source code:
+            /home-assistant-core/homeassistant/components/demo/config_flow.py
+            
+        Note that the "self.hass.data[DOMAIN][entry.entry_id]" object is present in the
+        OptionsFlow "async_step_init" step.  This allows you to access the client if one
+        is assigned to the data area.  All you need to do is assign a reference to the 
+        "entry:ConfigEntry" in the "__init__" in order to access it.  This saves you from
+        instantiating a new instance of the client to retrieve settings.  
+        """
+        errors: dict[str, str] = {}
+        
+        try:
+
+            # trace.
+            _logsi.EnterMethod(SILevel.Debug)
+            _logsi.LogDictionary(SILevel.Verbose, "'%s': OptionsFlow async_step_init is starting - user_input" % self._name, user_input)
+
+            # if not the initial entry, then save the entered options; otherwise, prepare the form.
+            if user_input is not None:
+            
+                # update config entry options from user input values.
+                self._Options[CONF_OPTION_DEVICE_DEFAULT] = user_input.get(CONF_OPTION_DEVICE_DEFAULT, None)
+                
+                # store the updated config entry options.
+                _logsi.LogDictionary(SILevel.Verbose, "'%s': OptionsFlow is updating configuration options - options" % self._name, self._Options)
+                return self.async_create_entry(
+                    title="", 
+                    data=self._Options
+                )
+
+            # load available spotify connect devices.
+            device_list:list[str] = await self.hass.async_add_executor_job(self._GetPlayerDevicesList)
+            if device_list is None:
+                errors["base"] = "no_player_devices"
+                return
+
+            # log device that is currently selected.
+            device_default:str = self._Options.get(CONF_OPTION_DEVICE_DEFAULT, None)
+            _logsi.LogVerbose("'%s': OptionsFlow option '%s' - SELECTED value: '%s'" % (self._name, CONF_OPTION_DEVICE_DEFAULT, device_default))
+                   
+            # create validation schema.
+            schema = vol.Schema(
+                {
+                    # note - DO NOT use "default" argument on the following - use "suggested_value" instead.
+                    # using "default=" does not allow a null value to be selected!
+                    vol.Optional(
+                        CONF_OPTION_DEVICE_DEFAULT,
+                        description={"suggested_value": self._Options.get(CONF_OPTION_DEVICE_DEFAULT)},
+                        ): SelectSelector(
+                                SelectSelectorConfig(
+                                    options=device_list,
+                                    mode=SelectSelectorMode.DROPDOWN
+                        )
+                    ),
+                }
+            )
+            
+            _logsi.LogVerbose("'%s': OptionsFlow is showing the init configuration options form" % self._name)
+            return self.async_show_form(
+                step_id="init", 
+                data_schema=schema, 
+                description_placeholders={CONF_NAME: self._name},
+                errors=errors or {}
+            )
+        
+        except Exception as ex:
+            
+            # trace.
+            _logsi.LogException(None, ex, logToSystemLogger=False)
+            raise
+        
+        finally:
+
+            # trace.
+            _logsi.LeaveMethod(SILevel.Debug)
+
+    
+    def _GetPlayerDevicesList(self) -> list:
+        """
+        Retrieves Spotify Connect device list from the Spotify Web API.
+        """
+        try:
+
+            # trace.
+            _logsi.EnterMethod(SILevel.Debug)
+            
+            # get configuration instance data so we can reference the client instance.
+            _logsi.LogVerbose("'%s': OptionsFlow is retrieving instance data" % self._name)
+            data:InstanceDataSpotifyPlus = self.hass.data[DOMAIN].get(self._entry.entry_id, None)
+            _logsi.LogObject(SILevel.Verbose, "'%s': OptionsFlow instance data object" % self._name, data)
+            
+            # get spotify connect player device list.
+            _logsi.LogVerbose("'%s': OptionsFlow is retrieving Spotify Connect player devices" % self._name)
+            devices:list[Device] = data.spotifyClient.GetPlayerDevices(refresh=True)
+            
+            # sort the results (in place) by Name, ascending order.
+            devices.sort(key=lambda x: (x.Name or "").lower(), reverse=False)
+           
+            # build string array of all devices.
+            result:list = []
+            item:Device
+            for item in devices:
+                result.append(item.SelectItemNameAndId)
+
+            # trace.
+            _logsi.LogArray(SILevel.Verbose, "'%s': OptionsFlow option '%s' - available values" % (self._name, CONF_OPTION_DEVICE_DEFAULT), result)
+            return result
+            
+        except Exception as ex:
+            
+            _logsi.LogError("'%s': OptionsFlow could not retrieve Spotify Connect player device list: %s" % (self._name, str(ex)))
+            return None
         
         finally: